refactor: :recycle: Remove meteor functions from loginViaQuery
