+++ apps/meteor/client/startup/index.ts
@@ -9,7 +9,6 @@ import './e2e';
 import './iframeCommands';
 import './incomingMessages';
 import './loadMissedMessages';
-import './loginViaQuery';
 import './messageObserve';
 import './messageTypes';
 import './reloadRoomAfterLogin';
