+++ apps/meteor/client/views/root/AppLayout.tsx
@@ -5,6 +5,7 @@ import PageLoading from './PageLoading';
 import { useCodeHighlight } from './hooks/useCodeHighlight';
 import { useEscapeKeyStroke } from './hooks/useEscapeKeyStroke';
 import { useGoogleTagManager } from './hooks/useGoogleTagManager';
+import { useLoginViaQuery } from './hooks/useLoginViaQuery';
 import { useMessageLinkClicks } from './hooks/useMessageLinkClicks';
 import { useSettingsOnLoadSiteUrl } from './hooks/useSettingsOnLoadSiteUrl';
 import { useWordPressOAuth } from './hooks/useWordPressOAuth';
@@ -57,6 +58,7 @@ const AppLayout = () => {
 	useCorsSSLConfig();
 	useAutoupdate();
 	useCodeHighlight();
+	useLoginViaQuery();
 
 	const layout = useSyncExternalStore(appLayout.subscribe, appLayout.getSnapshot);
